requirements.txt
streamlit==1.28.0
google-generativeai>=0.3.2
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.1
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0

